refactor cw20BalancesInfo to cw-core
