+++ packages/state/recoil/selectors/clients/cw-core.ts
@@ -1,5 +1,6 @@
-import { selectorFamily } from 'recoil'
+import { selectorFamily, waitForAll } from 'recoil'
 
+import { TokenInfoResponse } from '@dao-dao/types/contracts/cw20-gov'
 import { VotingModuleType, parseVotingModuleContractName } from '@dao-dao/utils'
 
 import { Cw20BaseSelectors, Cw20StakedBalanceVotingSelectors } from '.'
@@ -374,6 +375,40 @@ export const allCw20BalancesSelector = selectorFamily<
       }
 
       return balances
+    },
+})
+
+export const cw20BalancesInfoSelector = selectorFamily<
+  { symbol: string; denom: string; amount: string; decimals: number }[],
+  { address: string }
+>({
+  key: 'cw20BalancesInfo',
+  get:
+    ({ address }) =>
+    async ({ get }) => {
+      const cw20List =
+        get(allCw20BalancesSelector({ contractAddress: address })) ?? []
+
+      const cw20Info = get(
+        waitForAll(
+          cw20List.map(({ addr }) =>
+            Cw20BaseSelectors.tokenInfoSelector({
+              contractAddress: addr,
+              params: [],
+            })
+          )
+        )
+      ).filter(Boolean) as TokenInfoResponse[]
+
+      const cw20Tokens = cw20Info.map((info, idx) => {
+        return {
+          symbol: info.symbol,
+          denom: cw20List[idx].addr,
+          amount: cw20List[idx].balance,
+          decimals: info.decimals,
+        }
+      })
+      return cw20Tokens
     },
 })
