+++ packages/state/recoil/selectors/index.ts
@@ -3,4 +3,4 @@ export * from './chain'
 export * from './contract'
 export * from './price'
 export * from './proposal'
-export * from './tokens'
+export * from './pools'
